requirements.txt
# Core Web Framework
flask==2.3.3

# Data Processing & Analysis
pandas==2.2.2
numpy==1.26.4

# Machine Learning
scikit-learn==1.4.1.post1

# AI/ML Libraries
google-generativeai==0.4.1

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Additional dependencies that might be needed
# Jupyter notebooks (for AI_predictions modules)
jupyter==1.0.0
ipykernel==6.29.0

# Data visualization (optional)
matplotlib==3.8.2
seaborn==0.13.0

# Additional ML libraries (optional)
xgboost==2.0.3
lightgbm==4.1.0

# Development tools (optional)
python-dotenv==1.0.0 